Added AI based search aka "horizontal coding"

+++ qualcoder/ai_vectorstore.py
@@ -26,17 +26,22 @@ https://github.com/ccbogel/QualCoder
 https://qualcoder.wordpress.com/
 """
 
+import os
+os.environ["HF_HUB_DISABLE_TELEMETRY"] = "1" # for huggingface hub
+os.environ["ANONYMIZED_TELEMETRY"] = "0" # for chromadb
+
 from langchain.embeddings.sentence_transformer import SentenceTransformerEmbeddings
 from langchain.text_splitter import CharacterTextSplitter, RecursiveCharacterTextSplitter
 from langchain.vectorstores import Chroma
 from langchain.docstore.document import Document
+from huggingface_hub import hf_hub_url
+import requests
 from typing import (Any, Iterable, Optional, List)
-import os
 import logging
 import traceback
 from PyQt6 import QtWidgets
 from PyQt6.QtCore import QThreadPool
-import qualcoder.ai_async_worker
+from .ai_async_worker import Worker, AiException
 
 path = os.path.abspath(os.path.dirname(__file__))
 logger = logging.getLogger(__name__)
@@ -89,7 +94,7 @@ class E5SentenceTransformerEmbeddings(SentenceTransformerEmbeddings):
    
     
 class AiVectorstore():
-    """ This is the memory of AI research mate. 
+    """ This is the memory of AI buddy. 
     It manages a chromadb vectorstore with embeddings for all text based 
     source data in the project. This allows for semantic search and retrieval 
     of chunks of data which can then be further processed with a large 
@@ -102,11 +107,29 @@ class AiVectorstore():
     ready = False # If the chroma_db is busy ingesting documents, ready will be "False" since we cannot make any queries yet.   
     import_workers_count = 0
     # Setup the database 
-    model_name = "intfloat/multilingual-e5-small" # or intfloat/multilingual-e5-base
-    embedding_function = E5SentenceTransformerEmbeddings(model_name=model_name) 
+    model_name = "intfloat/multilingual-e5-large" # or intfloat/multilingual-e5-base
+    cache_folder = os.path.join(os.path.expanduser('~'), '.cache', 'torch', 'sentence_transformers')
+    model_folder = os.path.join(cache_folder, model_name.replace("/", "_"))
+    model_files = [
+        '1_Pooling/config.json',
+        '.gitattributes',
+        'config.json',
+        'modules.json',
+        'pytorch_model.bin',
+        'README.md',
+        'sentence_bert_config.json',
+        'sentencepiece.bpe.model',
+        'special_tokens_map.json',
+        'tokenizer.json',
+        'tokenizer_config.json'
+    ]
+    download_model_running = False
+    download_model_cancel = False
+    download_model_msg = ''
     chroma_db = None
     _is_closing = False
     collection_name = ''
+    embedding_function = None # E5SentenceTransformerEmbeddings(model_name=model_folder)
     
     def __init__(self, app, parent_text_edit, collection_name):
         self.app = app
@@ -114,15 +137,98 @@ class AiVectorstore():
         self.collection_name = collection_name
         self.threadpool = QThreadPool()
         self.threadpool.setMaxThreadCount(1)
+        
+    def embedding_model_is_cached(self) -> bool:
+        """Checks if the embeddings model from huggingface is already downloaded 
+        and in the local cache."""
+        return os.path.exists(os.path.join(self.model_folder, self.model_files[-1]))
+    
+    def _download_embedding_model(self, progress_callback=None):
+        """Downloads the embedding model to the local cache if necessary.
+
+        Args:
+            progress_callback (function(msg), optional): called regulary with an update
+                                                         message containing the file which
+                                                         is downloaded and the percent finished. 
+        """
+        if not self.embedding_model_is_cached():
+            self.download_model_running = True
+            self.download_model_cancel = False
+            for file_name in self.model_files:
+                local_path = os.path.join(self.model_folder, file_name)
+                if os.path.exists(local_path):
+                    continue # skipt this file, already downloaded
+                url = hf_hub_url(self.model_name, file_name)
+                
+                # create local dir if necessary
+                local_folder = os.path.dirname(local_path) # (may contain subdir to model_folder)
+                os.makedirs(local_folder, exist_ok=True)
+                tmp_filename = local_path + ".tmp"
+                
+                # download
+                r = requests.get(url, stream=True, timeout=20)
+                r.raise_for_status() # raise error on 404 etc.
+                with open(tmp_filename, "wb") as f:
+                    total_length = int(r.headers.get('content-length'))
+                    expected_size = (total_length/1024) + 1
+                    i = 0
+                    for chunk in r.iter_content(chunk_size = 1024):
+                        if self.download_model_cancel:
+                            return # cancel the download
+                        if chunk:
+                            f.write(chunk)
+                            i += 1
+                            msg = f'{os.path.basename(local_path)}: {round(i/expected_size * 100)}%'
+                            if progress_callback != None:
+                                progress_callback.emit(msg)
+                            print(msg, '                           ', end='\r', flush=True)
+                msg = f'{os.path.basename(local_path)}: 100%'
+                if progress_callback != None:
+                    progress_callback.emit(msg)
+                print(msg, '                           ', end='\r', flush=True)
+                if os.path.exists(local_path):
+                    os.remove(local_path)
+                os.rename(tmp_filename, local_path) 
+                
+    def _download_embedding_model_callback(self, msg):
+        # self.parent_text_edit.append(msg)
+        self.download_model_msg = msg
+    
+    def _download_embedding_model_finished(self):
+        if not self.ai_worker_running():
+            self.download_model_running = False
+            if self.download_model_cancel:
+                msg = _("AI buddy: Could not download all the necessary components, AI buddy will be disabled.")
+            else:
+                msg = _('AI buddy: Success, components downloaded and installed.')
+            self.parent_text_edit.append(msg)
+            logger.debug(msg)
+
+    def download_embedding_model(self):
+        """Downloads the embedding model to the local cache if necessary.
+        self.download_model_running will be True until all files have finished downloading.
+        """
+        self.download_model_running = True
+        worker = Worker(self._download_embedding_model) # Any other args, kwargs are passed to the run function
+        # worker.signals.result.connect()
+        worker.signals.finished.connect(self._download_embedding_model_finished)
+        worker.signals.progress.connect(self._download_embedding_model_callback)
+        worker.signals.error.connect(exception_handler)
+        self.threadpool.start(worker)
     
     def _open_db(self, progress_callback=None):
         if self._is_closing:
             return # abort when closing db
+        import sentence_transformers as st
+        self.st = st # this prevents the library from beeing garbage collected
         if self.app.project_path != '' and os.path.exists(self.app.project_path):
             db_path = self.app.project_path + '/vectorstore'
+            self.embedding_function = E5SentenceTransformerEmbeddings(model_name=self.model_folder)
+            collection_metadata = {"hnsw:space": "l2"} # {"hnsw:space": "cosine"} -> defines the distance function, cosine vs. Squared L2 (default). In my limited tests, l2 gives slightly better results, although cosine is usually recommended 
             self.chroma_db = Chroma(embedding_function=self.embedding_function, 
                                     persist_directory=db_path,
-                                    collection_name=self.collection_name)
+                                    collection_name=self.collection_name,
+                                    collection_metadata=collection_metadata)
         else:
             self.chroma_db = None
             raise FileNotFoundError(f'AI Vectorstore: project path "{self.app.project_path}" not found.')
@@ -136,9 +242,14 @@ class AiVectorstore():
             rebuild (bool, optional): Rebuild the vectorstore from the ground up. Defaults to False.
         """
         self._is_closing = False
-        self.ready = False
         
-        worker = qualcoder.ai_async_worker.Worker(self._open_db)  
+        self.parent_text_edit.append(_('Hi there! AI buddy is waking up...'))
+        QtWidgets.QApplication.processEvents() # update ui
+        
+        # import sentence_transformers # we have to import this here, not in the worker thread, or qualcoder will crash
+        #self.embedding_function = E5SentenceTransformerEmbeddings(model_name=self.model_folder)
+                              
+        worker = Worker(self._open_db)  
         if rebuild:
             worker.signals.finished.connect(self.rebuild_vectorstore)
         else:
@@ -150,13 +261,16 @@ class AiVectorstore():
         self.parent_text_edit.append(msg)
         
     def finished_import(self):
-        msg = _('AI research mate: Finished updating memory.')
         self.import_workers_count -= 1
         if self.import_workers_count <= 0:
             self.import_workers_count = 0
-            self.ready = True
+            msg = _("AI buddy: Ok, my memory is up to date.")
             self.parent_text_edit.append(msg)
             logger.debug(msg)
+
+        #if not self.ai_worker_running():
+        #    self.parent_text_edit.append(msg)
+        #    logger.debug(msg)
     
     def _import_document(self, id, name, text, update=False, progress_callback=None):
         if self._is_closing:
@@ -171,7 +285,6 @@ class AiVectorstore():
             else:
                 # skip the doc
                 return 
-        
         # add to chroma_db
         if self._is_closing:
             return # abort when closing db
@@ -179,7 +292,7 @@ class AiVectorstore():
         # split fulltext in smaller chunks 
         if text != '': # can only add embeddings if text is not empty
             if progress_callback != None:
-                progress_callback.emit(_('AI research mate: Memorizing document ') + f'"{name}"')
+                progress_callback.emit(_('AI buddy is memorizing ') + f'"{name}"')
 
             metadata = {'id': id, 'name': name}
             document = Document(page_content=text, metadata=metadata)
@@ -191,7 +304,7 @@ class AiVectorstore():
             # create embeddings for these chunks and store them in the chroma_db (with metadata)
             chunk_texts = [chunk.page_content for chunk in chunks]
             chunk_metadatas = [chunk.metadata for chunk in chunks]
-            self.chroma_db.add_texts(chunk_texts, chunk_metadatas)    
+            self.chroma_db.add_texts(chunk_texts, chunk_metadatas)  
 
     def import_document(self, id, name, text, update=False):
         """Imports a document into the chroma_db. 
@@ -199,14 +312,17 @@ class AiVectorstore():
         the chroma_db, it can be updated (update=True) 
         or skipped (update=False).
         This is an async process running in a background
-        thread. AiVectorstore.ready will be False
+        thread. AiVectorstore.is_ready() will return False
         until the import is finished.
 
         Args:
             id (integer): the database id
             update (bool, optional): defaults to False.
         """
-        worker = qualcoder.ai_async_worker.Worker(self._import_document, id, name, text, update) # Any other args, kwargs are passed to the run function
+        #self._import_document(id, name, text, update)
+        #return       
+        
+        worker = Worker(self._import_document, id, name, text, update) # Any other args, kwargs are passed to the run function
         # worker.signals.result.connect()
         worker.signals.finished.connect(self.finished_import)
         worker.signals.progress.connect(self.progress_import)
@@ -217,16 +333,22 @@ class AiVectorstore():
     def update_vectorstore(self):
         """Collects all text sources from the database and adds them to the chroma_db if 
         not already in there.  
-        """   
+        """
+        msg = _("AI buddy: Let's see if there are new documents...")
+        self.parent_text_edit.append(msg)
+        logger.debug(msg)   
         docs = self.app.get_file_texts()
         for doc in docs:
             self.import_document(doc['id'], doc['name'], doc['fulltext'], False)
             
     def rebuild_vectorstore(self):
-        logger.debug('AI research mate: rebuilding vectorstore')
+        msg = _('AI buddy: Give me some time to read all your documents...')
+        self.parent_text_edit.append(msg)
+        logger.debug(msg)
         # delete all the contents from the vectorstore
         ids = self.chroma_db.get(include=[])['ids']
-        self.chroma_db.delete(ids)
+        if len(ids) > 0:
+            self.chroma_db.delete(ids)
         # rebuild vectorstore
         self.update_vectorstore()
     
@@ -235,22 +357,26 @@ class AiVectorstore():
         from the vectorstore"""
         embeddings_list = self.chroma_db.get(where={"id" : id}, include=['metadatas'])
         if len(embeddings_list['ids']) > 0: # found it
-            self.parent_text_edit.append(f'AI buddy: forgetting "{embeddings_list["metadatas"][0]["name"]}"')
+            self.parent_text_edit.append("AI buddy: I'm gonna forget " + f'"{embeddings_list["metadatas"][0]["name"]}"')
             self.chroma_db.delete(embeddings_list['ids']) 
                
     def close(self):
         """Cancels the update process if running"""
         self._is_closing = True
+        self.download_model_cancel = True
         # cancel all waiting threads:
         self.threadpool.clear()
         self.threadpool.waitForDone(5000)
         self.chroma_db = None
-        self.ready = False
+        
+    def ai_worker_running(self):
+        return (self.import_workers_count > 0)
+        #return self.threadpool.activeThreadCount() > 0
             
     def is_ready(self) -> bool:
         """If the vectorstore is initiated and done importing data, 
         it is ready for queries."""
-        return self.chroma_db != None and self._is_closing == False and self.threadpool.activeThreadCount == 0
+        return self.chroma_db != None and self._is_closing == False and not self.ai_worker_running()
